# packages/contracts/openapi/taxmind-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateBatchData": {
        "additionalProperties": false,
        "properties": {
          "candidate_count": {
            "title": "Candidate Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "document_version_id": {
            "title": "Document Version Id",
            "type": "string"
          },
          "extraction_method": {
            "title": "Extraction Method",
            "type": "string"
          },
          "extractor_version": {
            "title": "Extractor Version",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_version_id",
          "extraction_method",
          "extractor_version",
          "model_name",
          "prompt_version",
          "status",
          "candidate_count",
          "created_at"
        ],
        "title": "CandidateBatchData",
        "type": "object"
      },
      "CandidateQueueResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCandidateData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "CandidateQueueResponse",
        "type": "object"
      },
      "CandidateReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "pattern": "^(approved|rejected)$",
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "decision"
        ],
        "title": "CandidateReviewRequest",
        "type": "object"
      },
...
      "CreatePublishBatchRequest": {
        "additionalProperties": false,
        "properties": {
          "candidate_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Candidate Ids",
            "type": "array"
          }
        },
        "required": [
          "candidate_ids"
        ],
        "title": "CreatePublishBatchRequest",
        "type": "object"
      },
      "CreatedCandidateBatchData": {
        "additionalProperties": false,
        "properties": {
          "batch": {
            "$ref": "#/components/schemas/CandidateBatchData"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCandidateData"
            },
            "title": "Candidates",
            "type": "array"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          }
        },
        "required": [
          "batch",
          "candidates",
          "created"
        ],
        "title": "CreatedCandidateBatchData",
        "type": "object"
      },
      "CreatedCandidateBatchResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreatedCandidateBatchData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "CreatedCandidateBatchResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeCandidateData": {
        "additionalProperties": false,
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "candidate_type": {
            "title": "Candidate Type",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "extraction_confidence": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Extraction Confidence",
            "type": "string"
          },
          "extraction_method": {
            "title": "Extraction Method",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "normalization_status": {
            "title": "Normalization Status",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "review_reason_safe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Reason Safe"
          },
          "review_status": {
            "title": "Review Status",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "source_chunk_id": {
            "title": "Source Chunk Id",
            "type": "string"
          },
          "source_document_id": {
            "title": "Source Document Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "batch_id",
          "candidate_type",
          "payload",
          "source_document_id",
          "source_chunk_id",
          "extraction_method",
          "extraction_confidence",
          "normalization_status",
          "review_status",
          "review_reason_safe",
          "reviewed_by",
          "reviewed_at",
          "created_at"
        ],
        "title": "KnowledgeCandidateData",
        "type": "object"
      },
      "KnowledgeCandidateResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/KnowledgeCandidateData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "KnowledgeCandidateResponse",
        "type": "object"
      },
      "KnowledgePublishBatchData": {
        "additionalProperties": false,
        "properties": {
          "approved_count": {
            "title": "Approved Count",
            "type": "integer"
          },
          "batch_type": {
            "title": "Batch Type",
            "type": "string"
          },
          "candidate_count": {
            "title": "Candidate Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "manifest_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Hash"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "rejected_count": {
            "title": "Rejected Count",
            "type": "integer"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "title": "Submitted At",
            "type": "string"
          },
          "validation_report": {
            "additionalProperties": true,
            "title": "Validation Report",
            "type": "object"
          }
        },
        "required": [
          "id",
          "batch_type",
          "scope",
          "status",
          "candidate_count",
          "approved_count",
          "rejected_count",
          "validation_report",
          "manifest_hash",
          "submitted_at",
          "published_at",
          "created_at"
        ],
        "title": "KnowledgePublishBatchData",
        "type": "object"
      },
      "KnowledgePublishBatchResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/KnowledgePublishBatchData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "KnowledgePublishBatchResponse",
        "type": "object"
      },
      "KnowledgeSnapshotData": {
        "additionalProperties": false,
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "manifest_hash": {
            "title": "Manifest Hash",
            "type": "string"
          },
          "pending_projection_event_count": {
            "title": "Pending Projection Event Count",
            "type": "integer"
          },
          "snapshot_code": {
            "title": "Snapshot Code",
            "type": "string"
          },
          "snapshot_type": {
            "title": "Snapshot Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "snapshot_code",
          "snapshot_type",
          "status",
          "manifest_hash",
          "activated_at",
          "created_at",
          "pending_projection_event_count"
        ],
        "title": "KnowledgeSnapshotData",
        "type": "object"
      },
      "KnowledgeSnapshotResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/KnowledgeSnapshotData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "KnowledgeSnapshotResponse",
        "type": "object"
      },
...
      "ResponseMeta": {
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TaxMind Pro",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/candidates": {
      "get": {
        "operationId": "list_pending_knowledge_candidates_api_v1_knowledge_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Pending Knowledge Candidates",
        "tags": [
          "knowledge-candidates"
        ]
      }
    },
    "/api/v1/knowledge/candidates/{candidate_id}/review": {
      "post": {
        "operationId": "review_knowledge_candidate_api_v1_knowledge_candidates__candidate_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeCandidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Knowledge Candidate",
        "tags": [
          "knowledge-candidates"
        ]
      }
    },
    "/api/v1/knowledge/document-versions/{version_id}/candidate-batches": {
      "post": {
        "operationId": "create_rule_based_candidate_batch_api_v1_knowledge_document_versions__version_id__candidate_batches_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedCandidateBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Rule Based Candidate Batch",
        "tags": [
          "knowledge-candidates"
        ]
      }
    },
...
    "/api/v1/knowledge/publish-batches": {
      "post": {
        "operationId": "create_knowledge_publish_batch_api_v1_knowledge_publish_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublishBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePublishBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Knowledge Publish Batch",
        "tags": [
          "knowledge-candidates"
        ]
      }
    },
    "/api/v1/knowledge/publish-batches/{batch_id}/materialize-snapshot": {
      "post": {
        "operationId": "materialize_knowledge_snapshot_api_v1_knowledge_publish_batches__batch_id__materialize_snapshot_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Materialize Knowledge Snapshot",
        "tags": [
          "knowledge-candidates"
        ]
      }
    },
    "/api/v1/knowledge/publish-batches/{batch_id}/validate": {
      "post": {
        "operationId": "validate_knowledge_publish_batch_api_v1_knowledge_publish_batches__batch_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePublishBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Knowledge Publish Batch",
        "tags": [
          "knowledge-candidates"
        ]
      }
    },
...
  }
}